+++ CHANGELOG.md
@@ -8,6 +8,39 @@ this format is adapted from [Keep a Changelog](https://keepachangelog.com/en/1.1
 ## [Unreleased]
 [Unreleased]: https://github.com/marbl/MetagenomeScope/compare/v1.2.0...HEAD
 
+### Added
+
+- Add the `-t`/`--vtsv` command-line option, to support reading
+  [Verkko](https://github.com/marbl/verkko)-style TSV files describing paths
+  ([#336](https://github.com/marbl/MetagenomeScope/issues/336)).
+
+- Various updates to the README.
+
+### Changed
+
+- Allow paths to span multiple connected components of the graph, since this
+  can occur in Verkko output.
+
+  - If a path spans multiple components, its `CC #` entry in the table of
+    available paths will be a comma-separated list of these components (rather
+    than just a single number). To account for this, the "type" of this column
+    has been changed from `number` to `text` -- this will mean that sorting
+    the paths table by this column will result in a different ordering.
+
+- Previously, we would create an extra copy of each path in a GFA file
+  representing the reverse-complementary copy of this path. Now, to be
+  consistent with other path filetypes and make interpretation easier, we
+  no longer do this
+  ([#357](https://github.com/marbl/MetagenomeScope/issues/357)).
+
+- Prevent node or edge IDs from beginning with `[N`. This makes parsing Verkko
+  TSV files easier: with this restriction, we can now unambiguously say if
+  something on a path is a gap or not.
+
+### Fixed
+
+- Cleaned up and added some more tests for the path-parsing parts of the code.
+
 
 ## [v1.2.0] - 2026-03-27 - Nonredundant drawing, bug fixes, and cleaning up
 [v1.2.0]: https://github.com/marbl/MetagenomeScope/releases/tag/v1.2.0